+++ engine/flink/management/periodic/src/main/scala/pl/touk/nussknacker/engine/management/periodic/model/PeriodicProcessDeployment.scala
@@ -9,8 +9,8 @@ import java.time.{Clock, LocalDateTime}
 // TODO: We should separate schedules concept from deployments - fully switch to ScheduleData and ScheduleDeploymentData
 case class PeriodicProcessDeployment(
     id: PeriodicProcessDeploymentId,
-    createdAt: LocalDateTime,
     periodicProcess: PeriodicProcess,
+    createdAt: LocalDateTime,
     runAt: LocalDateTime,
     scheduleName: ScheduleName,
     retriesLeft: Int,