+++ designer/server/src/main/scala/pl/touk/nussknacker/engine/CombinedProcessingTypeData.scala
@@ -1,7 +1,8 @@
 package pl.touk.nussknacker.engine
 
+import pl.touk.nussknacker.engine.definition.ComponentIdProvider
 import pl.touk.nussknacker.restmodel.process.ProcessingType
-import pl.touk.nussknacker.ui.component.{ComponentIdProvider, DefaultComponentIdProvider}
+import pl.touk.nussknacker.ui.component.ComponentIdProviderFactory
 import pl.touk.nussknacker.ui.process.{ProcessCategoryService, ProcessStateDefinitionService}
 import pl.touk.nussknacker.ui.process.ProcessStateDefinitionService.StatusNameToStateDefinitionsMapping
 
@@ -21,7 +22,7 @@ object CombinedProcessingTypeData {
         ProcessStateDefinitionService.createDefinitionsMappingUnsafe(processingTypes),
       // While creation of component id provider, we validate all component ids but fragments.
       // We assume that fragments cannot have overridden component id thus are not merged/deduplicated across processing types.
-      componentIdProvider = DefaultComponentIdProvider.createUnsafe(processingTypes, categoryService)
+      componentIdProvider = ComponentIdProviderFactory.createUnsafe(processingTypes, categoryService)
     )
   }