+++ docs/Changelog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 1.12.4 (?? Oct 2023)
-* [#4992](https://github.com/TouK/nussknacker/pull/4992) Fix: Sort PeriodicProcessDeployments by `runAt` and `createdAt` and instead of `runAt` only.
+* [#4992](https://github.com/TouK/nussknacker/pull/4992) Fix: List of periodic deployments is now sorted not only by schedule time but also by its creation time.
 
 1.12.3 (26 Oct 2023)
 1.12.2 (25 Oct 2023)