+++ docs/Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+1.13.0 (Not released yet)
+-------------------------
+* [#4711](https://github.com/TouK/nussknacker/pull/4711) Added AdditionalComponentsUIConfigProviderFactory API that allows changing components' configs without model reload
+
 1.12.4 (?? Oct 2023)
 * [#4992](https://github.com/TouK/nussknacker/pull/4992) Fix: List of periodic deployments is now sorted not only by schedule time but also by its creation time.